move TreeRep to data package
